Reestruturação
Nova estrutura de pastas,
Separação de Classes,
Uso de injeção de dependências
